refactor: bump scrypted-sdk to 0.1.1 and rename SDK glue modules
- Pin scrypted-sdk 0.1.1 in manifest.json and requirements_dev.txt
- Flatten deep scrypted_sdk.scrypted_python.scrypted_sdk.types imports to
  the package-root re-exports added in 0.1.1
- Rename rpc_transport.py -> sdk.py (SDK wiring) and client.py -> hub.py
  (per-entry lifecycle owner), resolving the naming collision with
  scrypted_sdk.client
