+++ custom_components/scrypted/climate.py
@@ -25,7 +25,7 @@ from .entity import (
     async_setup_scrypted_platform,
     device_matches,
 )
-from scrypted_sdk.scrypted_python.scrypted_sdk.types import ScryptedInterface
+from scrypted_sdk import ScryptedInterface
 
 SCRYPTED_TO_HVAC = {
     "Off": HVACMode.OFF,
