+++ custom_components/scrypted/cover.py
@@ -20,7 +20,7 @@ from .entity import (
     async_setup_scrypted_platform,
     device_matches,
 )
-from scrypted_sdk.scrypted_python.scrypted_sdk.types import ScryptedInterface
+from scrypted_sdk import ScryptedInterface
 
 COVER_DEVICE_CLASSES = {
     "Garage": CoverDeviceClass.GARAGE,
