+++ custom_components/scrypted/event.py
@@ -16,7 +16,7 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .const import SIGNAL_CONNECTION
 from .entity import ScryptedDeviceEntity, async_setup_scrypted_platform, device_matches
-from scrypted_sdk.scrypted_python.scrypted_sdk.types import ScryptedInterface
+from scrypted_sdk import ScryptedInterface
 
 _LOGGER = logging.getLogger(__name__)
