+++ custom_components/scrypted/light.py
@@ -24,7 +24,7 @@ from .entity import (
     async_setup_scrypted_platform,
     device_matches,
 )
-from scrypted_sdk.scrypted_python.scrypted_sdk.types import ScryptedInterface
+from scrypted_sdk import ScryptedInterface
 
 _LOGGER = logging.getLogger(__name__)
