+++ custom_components/scrypted/manifest.json
@@ -18,7 +18,7 @@
   "integration_type": "service",
   "iot_class": "local_push",
   "requirements": [
-    "scrypted-sdk==0.1.0"
+    "scrypted-sdk==0.1.1"
   ],
   "ssdp": [],
   "zeroconf": []
