+++ custom_components/scrypted/media_player.py
@@ -26,7 +26,7 @@ from homeassistant.helpers.entity import EntityDescription
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .entity import ScryptedDeviceEntity, async_setup_scrypted_platform, device_matches
-from scrypted_sdk.scrypted_python.scrypted_sdk.types import ScryptedInterface
+from scrypted_sdk import ScryptedInterface
 
 _LOGGER = logging.getLogger(__name__)
