+++ tests/test_hub.py
@@ -3,7 +3,7 @@ import pytest
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
-from custom_components.scrypted.client import ScryptedClient
+from custom_components.scrypted.hub import ScryptedClient
 from custom_components.scrypted.const import (
     DOMAIN,
     SIGNAL_DEVICE_UPDATE,
@@ -73,22 +73,22 @@ async def test_disconnect_triggers_reconnect(
     """A dropped connection reconnects with backoff after a failed attempt."""
     import asyncio
 
-    from custom_components.scrypted import client as client_module
+    from custom_components.scrypted import hub as hub_module
 
-    monkeypatch.setattr(client_module, "RECONNECT_INITIAL_DELAY", 0)
+    monkeypatch.setattr(hub_module, "RECONNECT_INITIAL_DELAY", 0)
     client = ScryptedClient(hass, entry)
     await client.async_connect()
 
     flaky_state = {"calls": 0}
-    real_connect = client_module.async_connect_sdk
+    real_connect = hub_module.async_connect_sdk
 
     async def flaky(*args, **kwargs):
         flaky_state["calls"] += 1
         if flaky_state["calls"] == 1:
-            raise client_module.ScryptedConnectionError("boom")
+            raise hub_module.ScryptedConnectionError("boom")
         return await real_connect(*args, **kwargs)
 
-    monkeypatch.setattr(client_module, "async_connect_sdk", flaky)
+    monkeypatch.setattr(hub_module, "async_connect_sdk", flaky)
 
     mock_connect_sdk.transport.handlers["disconnect"]()
     assert client.connected is False
